# perf: cache `Meta` directly and other improvements