--- a/frappe/tests/test_perf.py
+++ b/frappe/tests/test_perf.py
@@ -16,14 +16,21 @@ query. This test can be written like this.
 >>> 		get_controller("User")
 
 """
+import time
 import unittest
 
+from tenacity import retry, retry_if_exception_type, stop_after_attempt, wait_fixed
+
 import frappe
+from frappe.frappeclient import FrappeClient
 from frappe.model.base_document import get_controller
+from frappe.query_builder.utils import db_type_is
+from frappe.tests.test_query_builder import run_only_if
 from frappe.tests.utils import FrappeTestCase
 from frappe.website.path_resolver import PathResolver
 
 
+@run_only_if(db_type_is.MARIADB)
 class TestPerformance(FrappeTestCase):
 	def reset_request_specific_caches(self):
 		# To simulate close to request level of handling
@@ -33,6 +40,8 @@ class TestPerformance(FrappeTestCase):
 		frappe.clear_cache()
 
 	def setUp(self) -> None:
+		self.HOST = frappe.utils.get_site_url(frappe.local.site)
+
 		self.reset_request_specific_caches()
 
 	def test_meta_caching(self):
@@ -54,6 +63,36 @@ class TestPerformance(FrappeTestCase):
 
 		with self.assertQueryCount(0):
 			doc.get_invalid_links()
+
+	@retry(
+		retry=retry_if_exception_type(AssertionError),
+		stop=stop_after_attempt(3),
+		wait=wait_fixed(0.5),
+		reraise=True,
+	)
+	def test_req_per_seconds_basic(self):
+		"""Ideally should be ran against gunicorn worker, though I have not seen any difference
+		when using werkzeug's run_simple for synchronous requests."""
+
+		EXPECTED_RPS = 55  # measured on GHA
+		FAILURE_THREASHOLD = 0.1
+
+		req_count = 1000
+		client = FrappeClient(self.HOST, "Administrator", self.ADMIN_PASSWORD)
+
+		start = time.perf_counter()
+		for _ in range(req_count):
+			client.get_list("ToDo", limit_page_length=1)
+		end = time.perf_counter()
+
+		rps = req_count / (end - start)
+
+		print(f"Completed {req_count} in {end - start} @ {rps} requests per seconds")
+		self.assertGreaterEqual(
+			rps,
+			EXPECTED_RPS * (1 - FAILURE_THREASHOLD),
+			f"Possible performance regression in basic /api/Resource list  requests",
+		)
 
 	@unittest.skip("Not implemented")
 	def test_homepage_resolver(self):