## Main Improvement

`local.meta_cache` only lasted for the specific request, so it wasn't really useful. This has been removed completely, and replaced by `local.cache` (still per request). Instead of caching _serialized_ meta in Redis, the actual objects are now being cached.

### Before (Redis-only cache testing)
```python
In [10]: %timeit frappe.get_meta("Sales Invoice")
5.42 ms ± 58.3 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

### After (71% better in production!)
```python
In [4]: %timeit frappe.get_meta("Sales Invoice")
1.57 ms ± 43.4 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
 ```

## Other Improvements
   
- Initialize field map when creating meta, instead of inside `get_field`.
- Only translate required labels from default labels.
- `special_doctypes` is a set for faster lookup.

### Before

```python
In [9]: %timeit meta.get_field("localization")
104 ns ± 1.32 ns per loop (mean ± std. dev. of 7 runs, 10,000,000 loops each)

In [10]: %timeit meta.has_field("localization")
156 ns ± 2.12 ns per loop (mean ± std. dev. of 7 runs, 10,000,000 loops each)

In [11]: %timeit meta.get_label("name")
24.9 µs ± 156 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```

### After

```python
# 15% better
In [5]: %timeit meta.get_field("localization") 
88.9 ns ± 1.43 ns per loop (mean ± std. dev. of 7 runs, 10,000,000 loops each)

# 50% better
In [6]: %timeit meta.has_field("localization")
78.9 ns ± 0.515 ns per loop (mean ± std. dev. of 7 runs, 10,000,000 loops each)

# 90% better
In [8]: %timeit meta.get_label("name")
2.45 µs ± 20.2 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)

```