@ankush 2022-09-19
lgtm. 

Made few more changes for perf problems:
1. `get_doc` and `get_cached_doc` both now follow same caching logic = attempt pickling doc directly which works 99% of the time. `get_doc` was setting dict so far which made previous improvements kinda moot 🥲 
2. Removed `_allow_dict`, get_cached_value will now always fetch a document instead of asking for dict. Makes no difference now.